requirements.txt
spotipy==2.23.0
pandas==2.1.4
SQLAlchemy==2.0.23
python-dotenv==1.0.0
pytest==7.4.3
requests==2.31.0 